--- crypto/mlkem/ref/test/speed_print.h
@@ -1,9 +1,0 @@
-#ifndef PRINT_SPEED_H
-#define PRINT_SPEED_H
-
-#include <stddef.h>
-#include <stdint.h>
-
-void print_results(const char *s, uint64_t *t, size_t tlen);
-
-#endif
